create new simflow metadata module
